refactor: StylesTabのprop drilling解消とサブコンポーネント化
- PresetSelector, DisplaySettingsFormをstyles/に抽出
- StylesTab内でusePresetsを直接呼び出すように変更(20+ props -> 2 props)
- NewPresetModalをoptions.tsxからStylesTabに移動
- UsePresetsReturn型をexportに変更
- StylesTab: 541行 -> 146行

Closes #93
