+++ src/components/options/styles/index.ts
@@ -1,0 +1,2 @@
+export { PresetSelector } from './PresetSelector';
+export { DisplaySettingsForm } from './DisplaySettingsForm';
